Using C++11 and fix some warnings

+++ src/serialization.cpp
@@ -4,13 +4,13 @@
 #include <sstream>
 namespace zeroerr {
 
-IRObject* IRObject::alloc(unsigned size) {
+IRObject* IRObject::alloc(size_t size) {
     IRObject* list = new IRObject[size + 1];
     list[0].i      = size;
     return list + 1;
 }
 
-char* IRObject::alloc_str(unsigned size) {
+char* IRObject::alloc_str(size_t size) {
     char* s = new char[size + 1];
     s[size] = 0;
     return s;
